requirements.txt
pandas==2.0.3
plotly==5.9.0
scipy==1.11.1
Pillow==10.2.0
scikit_learn
streamlit==1.31.0
streamlit_option_menu==0.3.12
numpy
matplotlib

